loader: move host-specific functionality out to new host.go and netdev.go
